Added a basic navmesh movement controller, reworked the way plugin dependenies are loaded, Added some starter NPC Dialog, Added PlayfieldInit callback

+++ AOSharp.Bootstrap/Imports/MessageProtocol.cs
@@ -6,8 +6,8 @@ namespace AOSharp.Bootstrap.Imports
     public class MessageProtocol
     {
         [DllImport("MessageProtocol.dll", EntryPoint = "?DataBlockToMessage@@YAPAVMessage_t@@IPAX@Z", CallingConvention = CallingConvention.Cdecl)]
-        public static extern IntPtr DataBlockToMessage(int size, IntPtr pDataBlock);
+        public static extern int DataBlockToMessage(int size, IntPtr pDataBlock);
         [UnmanagedFunctionPointer(CallingConvention.Cdecl, CharSet = CharSet.Unicode, SetLastError = true)]
-        public delegate IntPtr DDataBlockToMessage(int size, IntPtr pDataBlock);
+        public delegate int DDataBlockToMessage(int size, IntPtr pDataBlock);
     }
 }
